refactor: use long double

+++ include/sculk/jsonc-c/jsonc.h
@@ -54,10 +54,10 @@ JSONC_API jsonc_variant_t jsonc_parse_content(const char* content, bool allow_tr
 
 JSONC_API int jsonc_get_variant_type(jsonc_variant_t handle);
 
-JSONC_API bool     jsonc_variant_to_bool(jsonc_variant_t handle);
-JSONC_API int64_t  jsonc_variant_to_signed(jsonc_variant_t handle);
-JSONC_API uint64_t jsonc_variant_to_unsigned(jsonc_variant_t handle);
-JSONC_API double   jsonc_variant_to_float(jsonc_variant_t handle);
+JSONC_API bool        jsonc_variant_to_bool(jsonc_variant_t handle);
+JSONC_API int64_t     jsonc_variant_to_signed(jsonc_variant_t handle);
+JSONC_API uint64_t    jsonc_variant_to_unsigned(jsonc_variant_t handle);
+JSONC_API long double jsonc_variant_to_float(jsonc_variant_t handle);
 
 JSONC_API const char* jsonc_variant_as_string(jsonc_variant_t handle);
 JSONC_API const char* jsonc_variant_as_big_int(jsonc_variant_t handle);
@@ -94,8 +94,8 @@ JSONC_API uint64_t       jsonc_object_get_unsigned(jsonc_object_t handle, const 
 JSONC_API void           jsonc_object_set_unsigned(jsonc_object_t handle, const char* key, uint64_t value);
 JSONC_API size_t         jsonc_object_get_any_int(jsonc_object_t handle, char* data, size_t length, const char* key);
 JSONC_API bool           jsonc_object_set_any_int(jsonc_object_t handle, const char* key, const char* value);
-JSONC_API double         jsonc_object_get_float(jsonc_object_t handle, const char* key);
-JSONC_API void           jsonc_object_set_float(jsonc_object_t handle, const char* key, double value);
+JSONC_API long double    jsonc_object_get_float(jsonc_object_t handle, const char* key);
+JSONC_API void           jsonc_object_set_float(jsonc_object_t handle, const char* key, long double value);
 JSONC_API size_t         jsonc_object_get_any_float(jsonc_object_t handle, char* data, size_t length, const char* key);
 JSONC_API bool           jsonc_object_set_any_float(jsonc_object_t handle, const char* key, const char* value);
 JSONC_API size_t         jsonc_object_get_any_number(jsonc_object_t handle, char* data, size_t length, const char* key);
@@ -150,8 +150,8 @@ JSONC_API size_t         jsonc_array_get_any_int(jsonc_array_t handle, char* dat
 JSONC_API bool           jsonc_array_set_any_int(jsonc_array_t handle, size_t index, const char* value);
 JSONC_API bool           jsonc_array_add_any_int(jsonc_array_t handle, const char* value);
 JSONC_API double         jsonc_array_get_float(jsonc_array_t handle, size_t index);
-JSONC_API void           jsonc_array_set_float(jsonc_array_t handle, size_t index, double value);
-JSONC_API void           jsonc_array_add_float(jsonc_array_t handle, double value);
+JSONC_API void           jsonc_array_set_float(jsonc_array_t handle, size_t index, long double value);
+JSONC_API void           jsonc_array_add_float(jsonc_array_t handle, long double value);
 JSONC_API size_t         jsonc_array_get_any_float(jsonc_array_t handle, char* data, size_t length, size_t index);
 JSONC_API bool           jsonc_array_set_any_float(jsonc_array_t handle, size_t index, const char* value);
 JSONC_API bool           jsonc_array_add_any_float(jsonc_array_t handle, const char* value);
